RPi/requirements.txt
colorzero==2.0
gpiozero==1.6.2
picamera==1.13
picamera2
opencv-python
numpy
#PyBluez==0.23
pyserial==3.5
requests~=2.27.1
